+++ StockPricing/src/main/java/com/stockmarket/helper/CompanyHelper.java
@@ -11,7 +11,7 @@ public class CompanyHelper {
 	private String companyCode;
 	private String companyName;
 	private String companyCEO;
-	private Integer companyTurnover;
+	private Long companyTurnover;
 	private String companyWebsite;
 	private List<StockDto> stockList;
 	private String stockExchange;
